+++ Kexa/models/export/azureBlobStorage/config.models.ts
@@ -3,7 +3,6 @@ import { SaveConfig } from "../config.models";
 export interface AzureBlobStorageSaveConfig extends SaveConfig {
     type: "azureBlobStorage";
     containerName: string;
-    blobName?: string;
     accountName?: string;
     accountKey?: string;
 }