+++ Kexa/models/export/mysql/config.models.ts
@@ -1,0 +1,5 @@
+import { SaveConfig } from "../config.models";
+
+export interface MySQLSaveConfig extends SaveConfig {
+    type: "mySQL";
+}